Move constant to its own file.
